Table Component Merge.
